Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,7 +1,7 @@
 
 int main(){
-int x;
-scanf("%.2d", &x);
-printf("You entered: %.2d", &x)
+float x;
+scanf("%f", &x);
+printf("You entered: %f", &x)
 return 0;
 }
